+++ app/models/event_file.rb
@@ -1,5 +1,8 @@
+# frozen_string_literal: true
+
+# Event PDF attachements
 class EventFile < Asset
-  mount_uploader :attachment, ProjectCaoUploader
+  mount_uploader :attachment, EventFileUploader
 
   validates :attachment, file_size: { maximum: 20.megabytes.to_i }
 end